# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "HelloCSE API",
        "description": "Documentation de l’API HelloCSE",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/profil": {
            "get": {
                "tags": [
                    "Profil"
                ],
                "summary": "Lister les profils",
                "description": "Retourne la liste de tous les profils",
                "operationId": "listProfils",
                "responses": {
                    "200": {
                        "description": "Liste des profils",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "nom": {
                                                "type": "string",
                                                "example": "Skywalker"
                                            },
                                            "prenom": {
                                                "type": "string",
                                                "example": "Luke"
                                            },
                                            "status": {
                                                "type": "string",
                                                "example": "actif ** n'est present que pour les utilisateurs authentifiés **"
                                            },
                                            "image_url": {
                                                "type": "string",
                                                "example": "/storage/profils/luke.jpg"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2025-04-17T14:30:00.000000Z"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2025-04-17T14:30:00.000000Z"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Profil"
                ],
                "summary": "Créer un profil",
                "description": "Crée un nouveau profil (authentification requise)",
                "operationId": "createProfil",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "nom",
                                    "prenom",
                                    "status"
                                ],
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "Skywalker"
                                    },
                                    "prenom": {
                                        "type": "string",
                                        "example": "Luke"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "actif",
                                            "inactif",
                                            "en attente"
                                        ],
                                        "example": "actif"
                                    },
                                    "image": {
                                        "description": "Image de profil (optionnelle)",
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Profil créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "profil": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "nom": {
                                                    "type": "string",
                                                    "example": "Skywalker"
                                                },
                                                "prenom": {
                                                    "type": "string",
                                                    "example": "Luke"
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "actif"
                                                },
                                                "administrateur_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "image_url": {
                                                    "type": "string",
                                                    "example": "profils/luke.jpg"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2025-04-17T14:30:00.000000Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2025-04-17T14:30:00.000000Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreurs de validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validation errors"
                                        },
                                        "errors": {
                                            "properties": {
                                                "nom": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "Le nom est requis."
                                                    }
                                                },
                                                "prenom": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "Le prénom est requis."
                                                    }
                                                },
                                                "status": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "Le statut est requis."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/profil/{id}": {
            "post": {
                "tags": [
                    "Profil"
                ],
                "summary": "Modifier un profil",
                "operationId": "editProfil",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du profil à modifier",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "Kenobi"
                                    },
                                    "prenom": {
                                        "type": "string",
                                        "example": "Ben"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "inactif"
                                    },
                                    "image": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profil modifié avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "profil": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "nom": {
                                                    "type": "string",
                                                    "example": "Kenobi"
                                                },
                                                "prenom": {
                                                    "type": "string",
                                                    "example": "Ben"
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "inactif"
                                                },
                                                "image_url": {
                                                    "type": "string",
                                                    "example": "profils/kenobi.jpg"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Profil non trouvé"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Profil"
                ],
                "summary": "Supprimer un profil",
                "operationId": "deleteProfil",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du profil à supprimer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Profil supprimé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Profil supprimé"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Profil non trouvé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}